+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/app.ts
@@ -8,15 +8,13 @@ import * as meta from './meta';
     selector: 'app',
     template: `
         <div style="width: 500px;">
-            <verticalbar [uiCls]="chartUiCls" [nodes]="chartNodes" [height]="chartHeight" [width]="chartWidth" [url]="chartUrl"></verticalbar>
+            <verticalbar [cls]="chartCls" [height]="chartHeight" [width]="chartWidth" [url]="chartUrl"></verticalbar>
         </div>
     `,
     directives: [Verticalbar]
 })
 export class App {
-    get chartNodes() {
-        return ["GREEN_KIRBY", "KLEIN_WATERS"];
-    }
+
     get chartHeight() {
         return window.innerHeight - 50;
     }
@@ -28,7 +26,8 @@ export class App {
     get chartUrl() {
         return meta.urls.data;
     }
-    get chartUiCls() {
+
+    get chartCls() {
         return "chart";
     }
 }