+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/verticalbar/verticalbar-d3-options.interface.ts
@@ -1,0 +1,7 @@
+import {Injectable} from 'angular2/core';
+
+@Injectable()
+export interface VerticalbarD3OptionsInterface {
+    width?:Number;
+    height?:Number;
+}