+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/verticalbar/verticalbar-d3-series.interface.ts
@@ -1,0 +1,8 @@
+import {Injectable} from 'angular2/core';
+
+@Injectable()
+export interface VerticalbarD3SeriesInterface {
+    x:Number;
+    y:Number;
+    label:String;
+}