+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/verticalbar/verticalbar-d3.interface.ts
@@ -1,7 +1,0 @@
-
-export interface VerticalbarD3Interface {
-    width?:number;
-    height?:number;
-
-
-}