+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/verticalbar/verticalbar.ts
@@ -1,83 +1,54 @@
-import {Component, OnInit} from 'angular2/core';
+import {Component, OnInit, Input} from 'angular2/core';
 import {Http, Headers, HTTP_PROVIDERS} from 'angular2/http';
 import {VerticalbarService} from './verticalbar.service';
 import {VerticalbarD3} from './verticalbar-d3';
-//import {VerticalBarOptionsInterface} from './verticalbar-options.interface';
+import {VerticalbarD3OptionsInterface} from './verticalbar-d3-options.interface';
+import {VerticalbarD3SeriesInterface} from './verticalbar-d3-series.interface';
 
 declare let d3:any;
 
 
 @Component({
     selector: 'verticalbar',
-    template: `<div class="{{uiCls}}-scrollbox" style="width: {{width}}px; height: {{height}}px;">
-                <verticalbar-d3 [options]="chartOptions" [data]="data"></verticalbar-d3>
+    template: `<div class="{{cls}}-scrollbox" style="width: {{width}}px; height: {{height}}px;">
+                <verticalbar-d3 [options]="options" [data]="data"></verticalbar-d3>
                </div>`,
     styleUrls: [
-      'src/verticalbar/verticalbar.css'
+        'src/verticalbar/verticalbar.css'
     ],
-    inputs: ['uiCls', 'height', 'width', 'nodes', 'url'],
     directives: [VerticalbarD3],
     providers: [HTTP_PROVIDERS, VerticalbarService]
 })
 export class Verticalbar implements OnInit {
-    data:any[];
-    height:Number;
-    width:Number;
-    nodes:String[];
-    url:String;
-    uiCls:String;
+    @Input() height:Number;
+    @Input() width:Number;
+    @Input() url:String;
+    @Input() cls:String;
+
+    private _data:Array<VerticalbarD3SeriesInterface>;
+
+
     constructor(private verticalBarService:VerticalbarService) {
-        this.data = [];
     }
+
     ngOnInit() {
-        let me = this;
         this.verticalBarService.request(this.url).subscribe((res:any) => {
-            var store = [];
-            if (res.children && this.nodes.length) {
-                this.nodes.map(function (node) {
-                    var titleNodes = res.children.filter(function (child:any) {
-                        return child.title == node;
-                    });
-                    if (titleNodes.length) {
-                        titleNodes.map(function (titleNode:any) {
-                            if (!titleNode.leaf) {
-                                titleNode.children.map(function (child:any) {
-                                    store.push(child);
-                                });
-                            }
-                        }, store);
-                    }
-                }, store);
+            if (res.items) {
+                this._data = res.items;
             }
-            me.bind(store);
         });
     }
-    bind(data) {
-        let root = {
-            key: this.nodes,
-            values: []
-        };
-        data.sort(function (a:any, b:any) {
-            return parseFloat(a.checking) - parseFloat(b.checking);
-        });
-        data.map(function (item:any) {
-            var namePieces = item.name.split(' '),
-                lastName = namePieces && namePieces.length > 1 ? namePieces[1].toUpperCase() : item.name.toUpperCase();
-            root.values.push({
-                "label": lastName,
-                "color": parseFloat(item.checking) < 0 ? "red" : "green",
-                "value": item.latitude,
-                "elementClick": function() {
-                    alert(item.checking);
-                }
-            });
-        }, root);
-        this.data.push(root);
+
+    get data():Array<VerticalbarD3SeriesInterface> {
+        return this._data;
     }
-    get chartOptions() {
+
+    get options():VerticalbarD3OptionsInterface {
         return {
             width: this.width,
             height: this.height
         }
     }
+
+
 }