+++ fiddles/angular2/fiddle-0031-PlayerToolbar/src/meta.ts
@@ -1,0 +1,9 @@
+
+var metadata = {
+    fiddleHeader: 'Angular 2 - Player Toolbar',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0031-PlayerToolbar',
+        data: "https://s3-us-west-2.amazonaws.com/s.cdpn.io/297733/data.json"
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};